requirements.txt
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.2
djangorestframework==3.3.3
gunicorn==19.4.5
psycopg2==2.6.1
drf-nested-routers==0.11.1
whitenoise==2.0.6
selenium==2.53.0
python-dateutil==2.5.3
